+++ WinthropInternship/src/AppBundle/Controller/FacultyLiaisonListController.php
@@ -22,13 +22,18 @@ class FacultyLiaisonListController extends Controller
      */
     public function indexAction()
     {
-        $em = $this->getDoctrine()->getManager();
-
-        $facultyLiaisonLists = $em->getRepository('AppBundle:FacultyLiaisonList')->findAll();
-
-        return $this->render('facultyliaisonlist/index.html.twig', array(
-            'facultyLiaisonLists' => $facultyLiaisonLists,
-        ));
+        if ($this->get('security.authorization_checker')->isGranted('ROLE_SUPER_ADMIN')) {
+            $em = $this->getDoctrine()->getManager();
+    
+            $facultyLiaisonLists = $em->getRepository('AppBundle:FacultyLiaisonList')->findAll();
+    
+            return $this->render('facultyliaisonlist/index.html.twig', array(
+                'facultyLiaisonLists' => $facultyLiaisonLists,
+            ));
+        
+        }else{
+            return $this->redirectToRoute('studentformone_index');
+        }
     }
 
     /**
@@ -39,22 +44,28 @@ class FacultyLiaisonListController extends Controller
      */
     public function newAction(Request $request)
     {
-        $facultyLiaisonList = new Facultyliaisonlist();
-        $form = $this->createForm('AppBundle\Form\FacultyLiaisonListType', $facultyLiaisonList);
-        $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $em->persist($facultyLiaisonList);
-            $em->flush();
-
-            return $this->redirectToRoute('facultyliaisonlist_show', array('id' => $facultyLiaisonList->getId()));
+        if ($this->get('security.authorization_checker')->isGranted('ROLE_SUPER_ADMIN')) {
+            $facultyLiaisonList = new Facultyliaisonlist();
+            $form = $this->createForm('AppBundle\Form\FacultyLiaisonListType', $facultyLiaisonList);
+            $form->handleRequest($request);
+    
+            if ($form->isSubmitted() && $form->isValid()) {
+                $em = $this->getDoctrine()->getManager();
+                $em->persist($facultyLiaisonList);
+                $em->flush();
+    
+                return $this->redirectToRoute('facultyliaisonlist_show', array('id' => $facultyLiaisonList->getId()));
+            }
+    
+            return $this->render('facultyliaisonlist/new.html.twig', array(
+                'facultyLiaisonList' => $facultyLiaisonList,
+                'form' => $form->createView(),
+            ));
+        
+        }else{
+            return $this->redirectToRoute('studentformone_index');
         }
-
-        return $this->render('facultyliaisonlist/new.html.twig', array(
-            'facultyLiaisonList' => $facultyLiaisonList,
-            'form' => $form->createView(),
-        ));
     }
 
     /**
@@ -65,12 +76,18 @@ class FacultyLiaisonListController extends Controller
      */
     public function showAction(FacultyLiaisonList $facultyLiaisonList)
     {
-        $deleteForm = $this->createDeleteForm($facultyLiaisonList);
+        if ($this->get('security.authorization_checker')->isGranted('ROLE_SUPER_ADMIN')) {
 
-        return $this->render('facultyliaisonlist/show.html.twig', array(
-            'facultyLiaisonList' => $facultyLiaisonList,
-            'delete_form' => $deleteForm->createView(),
-        ));
+            $deleteForm = $this->createDeleteForm($facultyLiaisonList);
+    
+            return $this->render('facultyliaisonlist/show.html.twig', array(
+                'facultyLiaisonList' => $facultyLiaisonList,
+                'delete_form' => $deleteForm->createView(),
+            ));
+        
+        }else{
+            return $this->redirectToRoute('studentformone_index');
+        }
     }
 
     /**
@@ -81,21 +98,27 @@ class FacultyLiaisonListController extends Controller
      */
     public function editAction(Request $request, FacultyLiaisonList $facultyLiaisonList)
     {
-        $deleteForm = $this->createDeleteForm($facultyLiaisonList);
-        $editForm = $this->createForm('AppBundle\Form\FacultyLiaisonListType', $facultyLiaisonList);
-        $editForm->handleRequest($request);
 
-        if ($editForm->isSubmitted() && $editForm->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+        if ($this->get('security.authorization_checker')->isGranted('ROLE_SUPER_ADMIN')) {
 
-            return $this->redirectToRoute('facultyliaisonlist_edit', array('id' => $facultyLiaisonList->getId()));
-        }
-
-        return $this->render('facultyliaisonlist/edit.html.twig', array(
-            'facultyLiaisonList' => $facultyLiaisonList,
-            'edit_form' => $editForm->createView(),
-            'delete_form' => $deleteForm->createView(),
-        ));
+            $deleteForm = $this->createDeleteForm($facultyLiaisonList);
+            $editForm = $this->createForm('AppBundle\Form\FacultyLiaisonListType', $facultyLiaisonList);
+            $editForm->handleRequest($request);
+    
+            if ($editForm->isSubmitted() && $editForm->isValid()) {
+                $this->getDoctrine()->getManager()->flush();
+    
+                return $this->redirectToRoute('facultyliaisonlist_edit', array('id' => $facultyLiaisonList->getId()));
+            }
+    
+            return $this->render('facultyliaisonlist/edit.html.twig', array(
+                'facultyLiaisonList' => $facultyLiaisonList,
+                'edit_form' => $editForm->createView(),
+                'delete_form' => $deleteForm->createView(),
+            ));
+        }else{
+            return $this->redirectToRoute('studentformone_index');
+        }        
     }
 
     /**
@@ -106,16 +129,23 @@ class FacultyLiaisonListController extends Controller
      */
     public function deleteAction(Request $request, FacultyLiaisonList $facultyLiaisonList)
     {
-        $form = $this->createDeleteForm($facultyLiaisonList);
-        $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $em->remove($facultyLiaisonList);
-            $em->flush();
+        if ($this->get('security.authorization_checker')->isGranted('ROLE_SUPER_ADMIN')) {
+
+            $form = $this->createDeleteForm($facultyLiaisonList);
+            $form->handleRequest($request);
+    
+            if ($form->isSubmitted() && $form->isValid()) {
+                $em = $this->getDoctrine()->getManager();
+                $em->remove($facultyLiaisonList);
+                $em->flush();
+            }
+    
+            return $this->redirectToRoute('facultyliaisonlist_index');
+        
+        }else{
+            return $this->redirectToRoute('studentformone_index');
         }
-
-        return $this->redirectToRoute('facultyliaisonlist_index');
     }
 
     /**
